refactor(git): split command families
